Added print stylesheet to isolate output on print
Left the preview when viewing in the browser. Discovered potential issue where light text such as the tickmarks and contact log headings are made much darker (too dark to write on) if 'print background images' is deselected in the print dialogue. If Eclipse terminals don't save the state of this option after restart then I'll have to remove the light text in favour of something else.