Added .jshintrc
Hidden (dotfile) config for JSHint, removes requirement for `/*global [var] */`
in every JS file which references document, window, Vue.js variables
etc.

Also added ES6 flag for future commit including Babel.js, which
will allow me to use template literals for Vue components.

I guess technically JSHint isn't included anywhere in the project,
but it's a common plugin for editors and this stops me having to
rewrite it (if I clone to another machine) or lose it (because
it's a dotfile). And I'm already including .gitignore, so what's
the difference, right?