Added example email to Microsoft account
Added example email address as the placeholder text in the Microsoft
Account section, as per #35. The placeholder is comprised of the first
and surnames converted to lowercase and separated by a `.`, suffixed
with the last 2 digits of the telephone number.

Currently the placeholder is not reflected in the output – the output
field is still left blank.